Add APIs for nForum announcement and discussion and author contributions
This commit is a conglomeration of changes which have already been
made on the live server.

1) We tighten up a certain spam pattern to try to close loopholes that
a certain user has been persistent in finding.

2) We fix a bug (null pointer exception) in the 500 error page.

3) We introduce an API for announcing non-trivial nLab page edits and
nLab page creations to the nForum. It is a Python API, which is
compiled to C using Cython, and then to a binary. This binary is
called from Instiki (the relevant changes are in wiki_controller.rb,
edit.rhtml, and new.rhtml.

4) We introduce an API for detecting whether an nForum discussion with
the same title (case insensitive) as a given nLab page exists, and
returning the link to the last comment of it if so. It is a Python API
of the same kind as that in 3).  We add a 'Discuss this page' item to
the menu on (almost) every nLab page to link (using this API) to the
relevant nForum discussion if it exists, and to the nForum home page
if it does not exist.

5) We introduce an API for listing author contributions to the nLab,
as well as an author's last n contributions for some n. We add a new
endpoint at /author for presenting this data, and link to it in a
couple of relevant places in the nLab (at the bottom of every nLab
page where author information is given; and in the menu of the nLab
page of an author, if such a page exists. Rails does not handle by
default . in a link in its routing, so we add something to
config/routes.rb to get around this.