Handle change of page name in nForum announcer
Previously, when the name of an nLab page was changed, this would lead
to a new discussion thread being created for the page at the nForum by
the announcer API. We now tweak the latter API in such a way that one
can indicate that the page has changed name; and so that in that case,
we update the title of the nForum discussion thread rather than create
a new thread.

We make the necessary changes to the wiki_controller in Instiki to
detect when a page name change has occurred, and to call the announcer
API appropriately in that case. We enforce within Instiki that a
change in page name must be announced on the nForum, i.e. cannot be
designated a 'trivial' edit.