Remove author information from pages
Towards the bottom of every nLab page, there was previously a line
providing the name of the person who made the last edit to/created the
page, along with their IP address and a timestamp for the change. We
change this to provide only the timestamp and a link to the history of
the page. The purpose of this is to emphasise that a page may have had
many contributors, and that the last person to edit a page may not be
its author or main contributor.

In addition, we make the following changes.

1) We slightly tweak the formatting of timestamps, adding an 'at'
between the date and the time.

2) We slightly tweak the author information on the history page, to
link to the /author endpoint rather than the nLab page of the author.
The /author endpoint exists for all contributors, whereas a
contributor may not have an nLab page; and info about nLab
contributions, rather than general information about the contributor,
seems more relevant here.

3) We tweak the author information on a 'revision' page, again linking
to /author rather than the nLab page of the author, and using the same
structure as on the /show page (including in particular a link to the
history of the page).