Change spam prevention on file upload page
Instiki is using a rather old, and apparently not particularly widely
used, plugin "form-spam-protection" for 'honeypot' detection of
spambot filling out of forms, which has not been updated since 2007.
This introduces javascript and cookie dependencies, and was hopelessly
buggy for the file upload page.

We remove use of this plugin on the file upload page, replacing it
with a simple honeypot which does not require javascript or cookies;
should this honeypot prove to not be sufficiently enticing, we can
improve it later as need arises.

The plugin is used on other pages as well; we should aim to remove it
from those as well.

We also take the opportunity to try to improve the look of the file
upload and file deletion pages, making them in particular less
cramped.