Remove restrictions on file id in routing URLs
Previously, a regex requirement for the endpoints involving files has
been in place which prevented, for instance, Rails being correctly
able to route when a filename including spaces was used. In such a
case, the internal error and stacktrace appeared in the browser for
the user.

We now drop all such requirements when it comes to routing for the
file endpoints. Validation can be carried out elsewhere as needed.
Some already occurs in the file_link method in lib/url_generator.