Add API for generating a bib entry for an nLab page
We add an endpoint, /cite, for providing bib entries for use when
citing an nLab page. One, for citing the current version of the page,
is under /show/, and the other, for a citing a specific revision of a
page, is under /revision/.

We also add an item 'Cite' to the menu at the bottom of the current
version of an nLab page to link to the /cite endpoint under /show; and
similarly add an item 'Cite' to the menu at the bottom of a historical
version of an nLab page to link to the /cite endpoint under /revision.

The endpoints display both an ASCII bib entry (that most likely to be
relevant) and a unicode version. Since nLab page names can be in
unicode, and since there is no generic way to convert unicode to
ASCII, we simply hard-code a conversion for each unicode character
which currently occurs in some nLab page title, with the exception of
a single Chinese entry and a few Russian entries, for which we provide
a romanisation. This hard-coded conversion will need to be kept up to
date.

As with the other APIs added recently, the API itself is written in
Python and compiled to a binary via C, by means of Cython. It is
called from the controller corresponding to the new endpoint.

We also take the opportunity to update some instructions in the README
files for the other Python APIs added recently.