Replace form_spam_protection with simple honeypot
We extend the work begun a few commits ago, to now completely remove
use of the form_spam_protection plugin, and replace it with a simple
honeypot on all forms where this plugin was employed. The plugin was
causing numerous problems, including a cookie-related one which
prevented submission of page edits in some cases.

If the simple honeypot proves to be inadequate, we can upgrade it.
Logging of when the honeypot catches spam, but we omit this for the
moment.