Add API for listing all pages and page categories
Previously, the /list endpoint has been intended to list all pages in
a wiki, as well as all page categories. Because the code Instiki uses
for linking to a wiki page is very slow, the page at the /list took an
eternity to load (indeed it did not do so before the read timeout took
effect).

We now rework this functionality. As usual, we write a Python API,
actually two APIs, for fetching the names of all pages in a wiki, for
fetching all names of pages in a wiki in a given category, for
fetching all page categories in a given wiki, for determining whether
or not a given string defines a page category, and for determining
whether or not a given wiki has any page categories. We compile these
APIs to binaries via C by means of Cython, and then make a couple of
new endpoints which call them.

The first of these is at /all_pages and /all_pages/[category_name],
listing all pages in a wiki and all pages in a certain category in a
wiki respectively. The second is at /page_categories, which lists all
page categories in a wiki. The UIs for the two endpoints are
intertwined (/all_pages links to /page_categories, which links to
/all_pages/[category_name]). We hand-code the links to pages rather
than make use of Instiki's mechanism, to speed things up.

We also re-route the old /list and /list/[category_name] endpoints to
the /all_pages and /all_pages/[category_name] respectively.

Finally, we make a few tweaks so that the new endpoints also can be
used when a web is published.