Replace page renderer
This is a conglomerate commit of a large amount of work carried out
since July, mainly around the replacement of the old page rendering
done in Instiki with a new renderer built out a number of Python APIs.
The initial motivation for this was the slowness of the old renderer
on large pages: we now render the page once and for all after an edit,
and just load the rendered version thereafter, which speeds things up
considerably.

The new renderer also brings a greater degree of flexibility and
modularity to the codebase, which makes it easier to add new features.
One such that has been added is the use of LaTeX syntax for various
things, such as theorem environments, table of contents, tikz
diagrams, and more.

There have been many small tweaks and bug fixes too, which also are
included here.

A significant component needed by the new renderer is a 'sequential
queue' for handling asynchronously the rendering of pages dependent on
a given one that has been edited.

Work on the new renderer is not finished. The new renderer is only
used currently for instance on 'live' pages, not on history or diff
pages.

There is also a little code pertaining to an experimental tex-to-HTML
parser that is included with this commit.