Add a spam pattern and fix rendering bugs
We make the following changes.

1) We have been affected by spam from different IP addresses with
certain content; we try to block this by means of the spam filter.

2) Maruku was not rendering content inside a \begin{centre}
\end{centre} block, due to the fact that we were processing this to
become a div block before applying Maruku: Maruku does not seem to
render anything within a HTML block. We add a workaround for this.

3) The same 'gylph' ids were being used across different SVG diagrams
when rendering tikz diagram code, causing labels to be completely
wrong when more than one SVG diagram was used on a page. We fix this
by prefixing the id with the diagram id.

4) Add missing import (previously it was present transitively by means
of something which now has been removed).