Robustify xymatrix rendering and improve nowiki handling
We make some tweaks to the xymatrix diagram renderer to allow it to
handle correctly a wider range of cases: curved arrows for example. We
move from \xymatrix{...} to \begin{xymatrix} ... \end{xymatrix} to
make it easier to detect the diagram correctly when mathematics is
used inside it.

We also tweak the <nowiki>...</nowiki> renderer and the way it is
called in certain cases to allow Tikz and Xymatrix diagram code, as
well as centred code, to be able to be displayed verbatim.