Link to published web pages correctly
Previously, in the new renderer at least, we have been using /show/ in
all cases when constructing a link to another page from the markdown
syntax for a link. This does not work well for a published web,
leading to an 'access denied'-like page. Instead, /published/ should
be used. We now make this change.