Harden tex diagram renderer and expand javascript blacklist
We improve the security of diagram rendering in a few ways.

1) We add a comment that

openin_any = p

should be set in texmf.cmf, to prevent opening files outside of the
working directory (the only files in the working directory for tex
compilation should be harmless).

2) We add a timeout to the tex compilation process, to help prevent a
denial of service attack.

3) We expand the tex blacklist as an additional deterrent, making it
in particular harder to write to the working directory, and harder to
carry out a denial of service attack.

In addition, we increase the protection against a javascript injection
attack, by blocking use of javascript: and blocking use of all event
attributes (which we could find) in which javascript might be used.
This blacklisting approach is most likely a temporary solution, but is
better than nothing!