Fix rollback null pointer exception
When making a rollback, the edit screen currently does not have an
announcement box. This led to a null pointer exception when making the
nForum announcement for a rollback (viewed as a trivial edit). We now
fix this.

In addition, we remove an unused argument from a function and from
calls to that function.