Change nLab logo
We replace the old nLab SVG logo with a new SVG logo. In the course of
this, we simplify slightly how the logo is embedded in the default
nLab page layout.

We take care to ensure that the SVG itself is of the correct size,
without needing CSS in a surrounding element. This should avoid a
problem which occasionally occurred with the old logo, namely that the
rendering of a page would stop with the huge actual SVG logo on the
screen, before it was resized.

We also tweak the CSS for the menu box for searching the nLab, so that
the box is horizontally aligned with the remainder of the menu.

Finally, we replace the old nLab icon (used for example in the browser
tab) with a new one obtained from the new logo.