Tweak new style, tidy up, introduce imagefromfile
We tweak the new style for the nLab introduced by Jake to take into
account user feedback.

We make a few other small tweaks that have accumulated on the deployed
nLab server, mostly of an administrative nature (e.g. spam filter).
Instead of relying on binary versions of certain of the Python APIs
compiled via Cython, we now simply use the Python source directly
(this was prompted by an issue with a suddenly missing library on the
nLab server needed by the binaries, but is desirable for simplicity in
any case; the newer Python APIs are already called directly).

Finally, we introduce a new piece of syntax for uploading a file and
for vertical space.