Add spam detector API
The nLab has suffered from a series of spam edits to pages in which
the entire content is erased and replaced by spam, and where the title
is changed.

Previously, our only defence against spam has been a combination of a
simple honeypot and a spam filter. We now introduce an API which
carries out a little more sophisticated check: the similarity of the
edited content and the original content is compared, returning a value
between 0 and 1, and this value must be below a certain threshold
determined by various criteria (for now: whether the author has a
corresponding nForum user, and whether or not the title has been
changed).