Create page content web directory for equation referencing
There has previously been a little bug in the equation reference
parser which failed to create a directory for a particular web within
the page content root directory if it did not already exist. A JSON
file for the purpose of equation referencing may need to be created in
this web before any pages are stored, if the first page which is
edited contains equations.

We now fix this.