Tweak float right styling in image from file block
On a mobile phone (used vertically), right floated images with width
larger than the screen size were being positioned with right edge
against the right margin, with the left part of the image disappearing
off to the left (without being able to scroll to see it).

We now tweak the CSS and rendering of an image from file block to
position the left part of the image against the left margin on a
mobile phone (or when the width of the browser is less than a certain
max).

We also correct a bug which led to the font size in a list being 1px
larger to the font size of paragraphs.