Fix duplication of clip ids in svg for tikz diagrams
Otherwise there were issues when more than one tikz diagram was
present on a page which was using some part of tikz leading 'clip' to
be used as part of the id of a corresponding svg element. We had the
same issue with 'glyph' previously; we make the same kind of fix here.