Treat 'show' and 'published' equally in a published web
If a web is changed to published, any links to the web's pages which
had been written manually, e.g. on external web pages, to /show/ would
previously have broken. We now tweak things so as to treat /show in
the same way as /published for a published web: essentially, we simply
carry out an 'internal redirect' (we do not actually redirect in the
HTML sense, since application logic is required to detect whether a
web is published).

Previously, anybody with the web's password could have used /show as
if the web was not published; this is no longer possible, but this is
not a significant limitation, as one can still edit a page, view its
history, etc, by using a direct link (/edit, etc).