Link to discussion thread on edit page and tweak menus
Rather than linking generically to the nForum, we now link to the
discussion thread for a particular page in the submission instructions
when editing it, linking generically to the nForum only when such a
thread does not exist (and tweaking the wording of the instructions in
this case).

We also implement various suggested improvements to the menus: adding
a link to the discussion thread for the page in the bottom menu (it
was previously only in the top menu); renaming certain menu items for
greater clarity; and removing a couple of items.