Attempt to fix bug in deletion of cached pages
Certain unicode characters are unable to be used if one uses tries to
remove a file with rm on Linux, and probably the same issue affects
File.delete in Ruby. This can lead to a bug where the cache of a page
is not deleted. We attempt to fix this; it seems to work in some
cases, at least.

We also commit the latest copy of author_to_user.