Add referencing mechanism
We add a central 'bibliography' to the nLab (backed by the database)
which can added to in the same way as new pages etc are created. We
add \cite and \bibitem blocks to the renderer to enable use of this
bibliography on individual pages.

When creating a reference, we internally treat this as a special case
of page creation, which allows us to make use of the functionality of
the latter, rather than double it.

Much more functionality can be added, and currently only bibliography
entries of 'article' type can be rendered, but the foundations are in
place.