Add search functionality for bibliography
Currently this works by fetching all references in the backend upon
page load, and storing these in a variable in the frontend; these
hidden references can be searched dynamically in the frontend. We
ensure that after navigating away and then back to the page using, for
example, the Back button in the browser, any entered search text is
remembered.