Add arXiv and DOI to article reference rendering
In addition, make it possible to pass these as fields in a submitted
BibTex entry.

We also tweak the parsing of the author field to rely on BibTex's curly
bracket mechanism for indicating surname.

Finally, we add some validation on the fields of a BibTex entry when
submitting an article, and tweak the instructions for submitting a
reference.