Handle page names ending in ? in all pages list
When a page name ends in '?', it was previously not being escaped,
causing the link to fail to render (the ? being interpreted as the
start of a query parameter). We now fix this.

The same kind of bug does not appear to be present in other
situations, e.g. links to nLab pages, or search results, probably
because the Markdown renderer is involved in those cases and correctly
escapes it.