Workaround asynchronicity bug
For some reason, the Ruby Open3 implementation seems to wait on the
completion of other child processes of a spawned process if we wish to
check the exit status. Thus the asynchronous rendering of the other
pages affected by rendering of a certain page which is implemented in
renderer.py and the sequential queue API has previously been
essentially useless.

We now introduce a work-around: we add an option to renderer.py to not
render these other pages, and in Rails now call the script twice, once
to check that the actual page renders (without rendering the other
pages), and once as before, but in the latter case without checking
the exit code or reading anything from stdout/stderr.