Rework centre/center block handling
Previously, we had resorted to an awful hack to render the content of
a \begin{centre} ... \end{centre} or \begin{center} ... \end{center}
block. This led to HTML which did not parse in some cases (e.g. with
use of \begin{imagefromfile} ... \end{imagefromfile} inside it when a
caption was used).

We now rework things to handle the rendering of the contents of such
blocks correctly.