Fix ncatlab/nlab-deployment#45.
The problem was as follows.
Name change in page.revise was picked up in the after_save cache sweeper call for revisions.build (but not for current_revision.update_attributes, strangely) even before the page model was saved.
So we cannot rely on the after_save handler to expire the old page name.